requirements.txt
numpy
fastapi
uvicorn
firebase_admin
python-dotenv
gensim